+++ backend/files/apps.py
@@ -1,0 +1,6 @@
+# from django.apps import AppConfig
+
+
+# class RemoteGameConfig(AppConfig):
+#     default_auto_field = 'django.db.models.BigAutoField'
+#     name = 'remote_game'
