--- backend/files/backend/remote_game/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class RemoteGameConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'remote_game'
