--- backend/files/backend/remote_game/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('update/', views.update_game_state, name='update'),
-]
