+++ backend/files/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+# from . import views
+
+urlpatterns = [
+    # path('update/', views.update_game_state, name='update'),
+]
